Setima Refatoracao das pendências do projeto e correção de funções para manipulação com o banco de dados

+++ src/main/resources/FunctionsSql/fn_atualizar_emprestimo.sql
@@ -6,7 +6,7 @@ CREATE OR REPLACE FUNCTION fn_atualizar_emprestimo(
     p_devolucao_prevista TIMESTAMP DEFAULT NULL,
     p_devolvido_em TIMESTAMP DEFAULT NULL
 )
-RETURNS INTEGER
+RETURNS BIGINT
 LANGUAGE plpgsql
 AS $$
 DECLARE
